Update Chat.java
Zombie-implemented the methods.

Adding a functioning addPartcipant method requires the userAccount class to be polished, don't want to pull it from another branch aside from the main branch (i don't know the consequences)

Same problems with removeParticipant, there is currently no established way to communicate to the server which user to remove from that chat and how the Server will handle this, but it has been implemented client-side

sendMessage is not implemented yet.

loadConversationHistory is partially implemented. See comment in the code.