UserManagment and ITUser Updates.
UserManagment:
Removes setUpChatList() since this was
created to assign chatList objects to the created UserAccount Objects.
This wouldnt work because the constructor for chatList had an error so
since that is the case this method is no longer needed.
Rewrote the rewriteUserFile() to read from the List<UserAccount> instead
of userCredentials. This allows us to write back to the file whenever we
remove a user. and also maintains the structure we need.

ITUser:
Modified newUser() to no longer contain an account type since we only
have one IT User at the moment. In the future we can implement it again
but for simplicity and testing purposes it makes it better.