Update Chat.java
Removed sendMessage. Not under jurisdiction of Chat class.

Rewrote addParticipants. Simply adds UserAccount to data structure after it is validated in Chatroom, where the network handle is.

Rewrote removeParticipants. Checks if given username is in Chat participants. Returns false if not.

Removed loadConversationHistory. This file handling responsibility has been shifted to ConversationHistory.

Added a method, getConversationHistory, that returns the handle to the ConversationHistory of the Chat.